# openapi.json
{
  "components": {
    "schemas": {
      "APIResponse": {
        "description": "Top-level envelope returned by every endpoint (untyped ``data``).",
        "properties": {
          "data": {
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "APIResponse",
        "type": "object"
      },
...
      "AdminDropConfigPatch": {
        "additionalProperties": false,
        "description": "Admin logistics patch for a drop (omit = leave alone; explicit null = clear).\n\nWindow fields accept epoch-ms integers on the wire (same as GET detail).",
        "properties": {
          "applyCloseAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applycloseat"
          },
          "applyOpenAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applyopenat"
          },
          "campaignHashtag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaignhashtag"
          },
          "capacityTotal": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacitytotal"
          },
          "totalProductUnits": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalproductunits"
          }
        },
        "title": "AdminDropConfigPatch",
        "type": "object"
      },
...
      "BrandDropCreateRequest": {
        "description": "Body for ``POST /api/brands/me/drops`` (architecture \u00a78.4).",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description"
        ],
        "title": "BrandDropCreateRequest",
        "type": "object"
      },
...
      "DataResponse_AdminDropDetail_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminDropDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[AdminDropDetail]",
        "type": "object"
      },
...
      "DataResponse_ApplicationResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[ApplicationResponse]",
        "type": "object"
      },
...
      "DataResponse_BrandDropDetailResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandDropDetailResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[BrandDropDetailResponse]",
        "type": "object"
      },
      "DataResponse_BrandDropResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandDropResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[BrandDropResponse]",
        "type": "object"
      },
...
      "DataResponse_DropDetailResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DropDetailResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[DropDetailResponse]",
        "type": "object"
      },
...
      "DataResponse_OkResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OkResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[OkResponse]",
        "type": "object"
      },
...
      "DataResponse_list_AdminDropItem__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AdminDropItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[list[AdminDropItem]]",
        "type": "object"
      },
...
      "DataResponse_list_BrandDropListItem__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BrandDropListItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[list[BrandDropListItem]]",
        "type": "object"
      },
...
      "DataResponse_list_DropFeedItem__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DropFeedItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[list[DropFeedItem]]",
        "type": "object"
      },
...
      "DropApplyRequest": {
        "description": "Body for ``POST /api/drops/{id}/apply`` (architecture \u00a77.1).",
        "properties": {
          "pitch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pitch"
          }
        },
        "title": "DropApplyRequest",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "Machine-readable error payload returned in `APIResponse.error`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "NotifyRequest": {
        "description": "Body for ``POST /api/drops/{id}/notify`` (architecture \u00a77.1).",
        "properties": {
          "reminderMinutes": {
            "title": "Reminderminutes",
            "type": "integer"
          }
        },
        "required": [
          "reminderMinutes"
        ],
        "title": "NotifyRequest",
        "type": "object"
      },
      "OkResponse": {
        "description": "Generic success ack (logout, notify, unlink, dismiss, password flows).",
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Buzz API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/drops": {
      "get": {
        "operationId": "list_drops_endpoint_api_admin_drops_get",
        "parameters": [
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "in": "query",
            "name": "attention",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Attention"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_AdminDropItem__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Drops Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/drops/{drop_id}": {
      "get": {
        "operationId": "get_drop_detail_endpoint_api_admin_drops__drop_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_AdminDropDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Drop Detail Endpoint",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "patch_drop_config_endpoint_api_admin_drops__drop_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDropConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_AdminDropDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Patch Drop Config Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/brands/me/drops": {
      "get": {
        "operationId": "list_brand_drops_api_brands_me_drops_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_BrandDropListItem__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Brand Drops",
        "tags": [
          "brands"
        ]
      },
      "post": {
        "operationId": "create_drop_api_brands_me_drops_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandDropCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_BrandDropResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Drop",
        "tags": [
          "brands"
        ]
      }
    },
    "/api/brands/me/drops/{drop_id}": {
      "get": {
        "operationId": "get_brand_drop_detail_api_brands_me_drops__drop_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_BrandDropDetailResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Brand Drop Detail",
        "tags": [
          "brands"
        ]
      }
    },
...
    "/api/drops": {
      "get": {
        "description": "Org drop browse feed (JWT + ``org`` role + ``active``).",
        "operationId": "list_drops_api_drops_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_DropFeedItem__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Drops",
        "tags": [
          "drops"
        ]
      }
    },
    "/api/drops/{drop_id}": {
      "get": {
        "description": "Org-facing drop detail.",
        "operationId": "get_drop_api_drops__drop_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_DropDetailResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Drop",
        "tags": [
          "drops"
        ]
      }
    },
    "/api/drops/{drop_id}/apply": {
      "post": {
        "description": "Apply to a drop (``DROP_NOT_OPEN`` / ``ALREADY_APPLIED`` / ``CAPACITY_EXCEEDED``).",
        "operationId": "apply_drop_api_drops__drop_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DropApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_ApplicationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Apply Drop",
        "tags": [
          "drops"
        ]
      }
    },
    "/api/drops/{drop_id}/notify": {
      "delete": {
        "description": "Remove the caller org's reminder for a drop (idempotent).",
        "operationId": "clear_drop_notify_api_drops__drop_id__notify_delete",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Clear Drop Notify",
        "tags": [
          "drops"
        ]
      },
      "post": {
        "description": "Set/replace the caller org's reminder for a drop.",
        "operationId": "set_drop_notify_api_drops__drop_id__notify_post",
        "parameters": [
          {
            "in": "path",
            "name": "drop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Set Drop Notify",
        "tags": [
          "drops"
        ]
      }
    },
...
  }
}